pyproject.toml
[tool.poetry]
name = "structured_output_llm"
version = "0.1.0"
description = "Enforcing structured output in an LLM"
packages = [{ include = "structured_output_llm", from = "src" }]
authors = []

readme = "README.md"
include = [
    "README.md",
    "LICENSE.txt",
    "src/structured_output_llm/py.typed",
]
exclude = [
    "*.so",
    "*.pyc",
    "*~",
    "#*",
    ".git*",
    ".coverage*",
    "DS_Store",
    "__pycache__",
]


[tool.poetry.dependencies]
python = ">=3.10, <3.11"
outlines = "^0.0.8"
pydantic = "^1.10.9"
openai = "^0.28.1"
transformers = "^4.34.0"
guardrails-ai = "^0.2.4"

[tool.poetry.dev-dependencies]
# type hints
mypy = "1.3.0"
# formatting
black = "23.7.0"
# linting
ruff = "0.0.269"
# import sorting
isort = "5.12.0"
# testing
pytest = "7.3.1"
pytest-sugar = "0.9.7"
# on-demand environments
nox = "2023.4.22"
nox-poetry = "1.0.2"

[tool.mypy]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
warn_no_return = false
allow_untyped_defs = true
allow_untyped_decorators = true

[tool.isort]
profile = "black"
combine_as_imports = true
known_first_party = ["structured_output_llm"]
order_by_type = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
